Updates from Successful Users
This update includes a number of small adjustments made by two UNLV Ph.D. students from the Department of Political Science. K. Eugenis and K. Retzl adjusted the template to conform with new formatting rules. They have also worked out some of the bugs that have plagued this template. The adjustments are small, and are perhaps too numerous to mention.